# Devdocs: Fix page change lag in devdocs (fix #22756)