--- a/package.json
+++ b/package.json
@@ -128,6 +128,7 @@
     "react-day-picker": "6.2.1",
     "react-dom": "16.2.0",
     "react-hot-loader": "1.3.1",
+    "react-lazily-render": "1.0.1",
     "react-modal": "3.1.11",
     "react-pure-render": "1.0.2",
     "react-redux": "5.0.6",