This aims to fix #22756 by reducing the delay in page changes on the devdocs pages. Especially `/devdocs/blocks` and `/devdocs/design`.

------

Originally posted in #22756:

I talked to @nb about this but it seems the main issue is around the (very large) number of components mounted inside these pages (`/devdocs/blocks` and `/devdocs/design`). Pagination through the components wouldn't be a good UX so continuing to load all of the examples on the page is what I aimed to keep.

First off I tried using something like [`react-virtualized`](https://github.com/bvaughn/react-virtualized) to load only the content on screen but because the components are of varying heights that lead to only a ~10-20% reduction in mounting time and a worse scrolling UX. It also meant the entire page would never be mounted at once so in-browser "find" UX was ruined.

Using [`react-lazily-render`](https://github.com/jameslnewell/react-lazily-render) turned out to work much better, and using it means we can load the first X number of components immediately (basically enough to fill the viewport) so the component mounts in a fraction of the time it did in the past (for reference, it mounts faster using React development environment with the change than it currently does in production with the old "mount everything at once" approach). We can then load the rest of the components either: onScroll or after a few seconds with a timeout. The underlying code used is a bit gnarly but the API of the Devdocs `Collection` remains the same with the heavy-lifting/lazy-loading happening transparently, so this approach will scale if more components/pages are added.