@tofumatt 2018-03-07
I should point out that I tested this locally on my MacBook (mid-2017, 1.4 GHz Intel Core i7) in Firefox Developer Edition, Chrome release, and Safari on MacOS and saw what I felt to be a noticeable performance increase. I didn't script a performance benchmark test but running the profiler in Firefox and Chrome saw a least a full second decrease in page load time switching between pages.

This is obviously anecdotal so I'd love a more comprehensive perf test (I could script one if there aren't any easily accessible tools to do one) and/or other developers eyes on this one.

This is what a few tests in Firefox yielded:

### Current Devdocs

`render()` time (Promise callback) | Click + `render()` total time
---------------------------------- | -----------------------------
3,122ms                            | 3,385ms
1,821ms                            | 1,940ms
1,775ms                            | 1,915ms

### Lazy render implementation

`render()` time (Promise callback) | Click + `render()` total time
---------------------------------- | -----------------------------
539ms                              | 800ms
384ms                              | 600ms
300ms                              | 580ms