@tofumatt 2018-03-07
Regarding removing existing rendered elements–I tried that with `react-virtualized` but it didn't work great. Here are the issues:

1. example components are variable height, so you still need to render them off-screen to calculate height on initial render so the scrolling works well. This means initial render is actually not much faster–sometimes it saves `~300ms`, but adds a lot of complexity to the render
2. the scrolling experience is diminished then and scrolling itself becomes a lot slower because we're managing which components are rendered and which aren't based on scroll events. These are debounced so it actually takes some time to catch up to the scroll, making the page feel slow
3. removing elements from the page means in-browser find is not longer usable. This is a minor point.

----

Ultimately, this is sort of an infinite scroll solution but with an added background load of the rest of the content (which is known and not remotely-fetched) using the `setTimeout`. The result is clicking on the page feels fast and switch between pages is fast. But we aren't forcing the user to scroll many times to load all content... I actually tried that approach by wrapping *every* `<DocsExampleWrapper>` in `<LazyRender>` but it made scrolling down really annoying... I can attach a screencast if you like. This solution means provides two user stories, basically:

### User sees placeholder scrolling
1. Click "Blocks" or "Design"
2. Page loads quickly with first few elements
3. User instantly starts scrolling to bottom
4. Placeholders appear while rest of components are mounted (this takes about 1-2s)
5. Components are all mounted and page is fully ready

### User doesn't see placeholders
1. Click "Blocks" or "Design"
2. Page loads quickly with first few elements
3. User doesn't interact with page right away, reads the first few components, or otherwise does something else
4. `setTimeout` fires after 2s and loads rest of page in ~1-2s
5. Components are all mounted and page is fully ready, entirely transparent to the user but without the initial render lag

Hope that makes sense 😄