@tofumatt 2018-03-07
Hey @nb, I've made a few changes to how this works. The original was more like an infinite scroll in that it waited a few seconds before just automatically mounting the rest of the components. Turns out we could wait a lot less time and still get the desired effect, with an added benefit of the re-render being nearly totally transparent to most users.

I also removed the hack/proof-of-concept approach of digging into the child component's `setState` call from the parent and just made a `DelayRender` component that loads after a set amount of time.

This seems like an odd patch I'll admit looking at it, but the solution definitely improves responsiveness of the page. This new one feels less like infinite scroll as the component simply mounts the rest of the components after 20ms, but this makes a notable difference especially in dev mode.